chore: shift sp.oidc service to separate file
