+++ src/app/modules/submission/email-submission/email-submission.controller.ts
@@ -21,7 +21,8 @@ import {
 import { MyInfoService } from '../../myinfo/myinfo.service'
 import * as MyInfoUtil from '../../myinfo/myinfo.util'
 import { SgidService } from '../../sgid/sgid.service'
-import { SpcpService, SpOidcService } from '../../spcp/spcp.service'
+import { SpOidcService } from '../../spcp/sp.oidc.service'
+import { SpcpService } from '../../spcp/spcp.service'
 import * as EmailSubmissionMiddleware from '../email-submission/email-submission.middleware'
 import * as SubmissionService from '../submission.service'
 import { extractEmailConfirmationData } from '../submission.utils'
