+++ src/app/modules/submission/encrypt-submission/encrypt-submission.controller.ts
@@ -26,7 +26,8 @@ import { MalformedParametersError } from '../../core/core.errors'
 import { ControllerHandler } from '../../core/core.types'
 import { PermissionLevel } from '../../form/admin-form/admin-form.types'
 import * as FormService from '../../form/form.service'
-import { SpcpService, SpOidcService } from '../../spcp/spcp.service'
+import { SpOidcService } from '../../spcp/sp.oidc.service'
+import { SpcpService } from '../../spcp/spcp.service'
 import { getPopulatedUserById } from '../../user/user.service'
 import * as VerifiedContentService from '../../verified-content/verified-content.service'
 import { WebhookFactory } from '../../webhook/webhook.factory'
